6338 backend search combo (#1130)
* needed changes for auto-analyse to run


* merged xpro/bc endpoints together, added options for desgnations / name structure, lint fixes


* name-analysis update for new flow


* new designation rule


* version upgrade
