+++ api/namex/VERSION.py
@@ -1,1 +1,1 @@
-__version__='1.0.22'
+__version__='1.1.0'
