--- api/namex/resources/auto_analyse/paths/bc_name_analysis/__init__.py
@@ -1,2 +1,0 @@
-from .bc_name_analysis import BcNameAnalysis
-from .bc_name_analysis import api as bc_name_analysis_api
