--- api/namex/resources/auto_analyse/paths/bc_name_analysis/bc_name_analysis_issues.py
@@ -1,1 +1,0 @@
-from ...issues import *
