--- api/namex/resources/auto_analyse/paths/bc_name_analysis/issues/__init__.py
@@ -1,18 +1,0 @@
-from .check_is_valid import BcCheckIsValidIssue
-from .add_descriptive_word import BcAddDescriptiveWordIssue
-from .add_distinctive_word import BcAddDistinctiveWordIssue
-from .check_is_valid import BcCheckIsValidIssue as CheckIsValidIssue
-from .contains_unclassifiable_word import BcContainsUnclassifiableWordIssue
-from .contains_words_to_avoid import BcContainsWordsToAvoidIssue
-from .corporate_name_conflict import BcCorporateNameConflictIssue
-from .queue_name_conflict import BcQueueNameConflictIssue
-from .designation_mismatch import BcDesignationMismatchIssue
-from .designation_removal import  BcDesignationRemovalIssue
-from .designation_misplaced import BcDesignationMisplacedIssue
-from .designation_non_existent import BcDesignationNonExistentIssue
-from .end_designation_more_than_once import BcEndDesignationMoreThanOnceIssue
-from .incorrect_category import BcIncorrectCategoryIssue as BcIncorrectCategoryIssue
-from .name_requires_consent import BcNameRequiresConsentIssue
-from .too_many_words import BcTooManyWordsIssue
-from .word_special_use import BcWordSpecialUseIssue as BcWordSpecialUseIssue
-from .incorrect_year import BcIncorrectYearIssue
