--- api/namex/resources/auto_analyse/paths/bc_name_analysis/issues/add_descriptive_word.py
@@ -1,5 +1,0 @@
-from namex.resources.auto_analyse.issues import AddDescriptiveWordIssue
-
-
-class BcAddDescriptiveWordIssue(AddDescriptiveWordIssue):
-    pass
