--- api/namex/resources/auto_analyse/paths/bc_name_analysis/issues/add_distinctive_word.py
@@ -1,5 +1,0 @@
-from namex.resources.auto_analyse.issues import AddDistinctiveWordIssue
-
-
-class BcAddDistinctiveWordIssue(AddDistinctiveWordIssue):
-    pass
