--- api/namex/resources/auto_analyse/paths/bc_name_analysis/issues/check_is_valid.py
@@ -1,5 +1,0 @@
-from namex.resources.auto_analyse.issues import CheckIsValidIssue
-
-
-class BcCheckIsValidIssue(CheckIsValidIssue):
-    pass
