--- api/namex/resources/auto_analyse/paths/bc_name_analysis/issues/contains_unclassifiable_word.py
@@ -1,5 +1,0 @@
-from namex.resources.auto_analyse.issues import ContainsUnclassifiableWordIssue
-
-
-class BcContainsUnclassifiableWordIssue(ContainsUnclassifiableWordIssue):
-    pass
