--- api/namex/resources/auto_analyse/paths/bc_name_analysis/issues/contains_words_to_avoid.py
@@ -1,5 +1,0 @@
-from namex.resources.auto_analyse.issues import ContainsWordsToAvoidIssue
-
-
-class BcContainsWordsToAvoidIssue(ContainsWordsToAvoidIssue):
-    pass
