--- api/namex/resources/auto_analyse/paths/bc_name_analysis/issues/corporate_name_conflict.py
@@ -1,5 +1,0 @@
-from namex.resources.auto_analyse.issues import CorporateNameConflictIssue
-
-
-class BcCorporateNameConflictIssue(CorporateNameConflictIssue):
-    pass
