--- api/namex/resources/auto_analyse/paths/bc_name_analysis/issues/designation_misplaced.py
@@ -1,5 +1,0 @@
-from namex.resources.auto_analyse.issues import DesignationMisplacedIssue
-
-
-class BcDesignationMisplacedIssue(DesignationMisplacedIssue):
-    pass
