--- api/namex/resources/auto_analyse/paths/bc_name_analysis/issues/designation_non_existent.py
@@ -1,5 +1,0 @@
-from namex.resources.auto_analyse.issues import DesignationNonExistentIssue
-
-
-class BcDesignationNonExistentIssue(DesignationNonExistentIssue):
-    pass
