--- api/namex/resources/auto_analyse/paths/bc_name_analysis/issues/incorrect_category.py
@@ -1,5 +1,0 @@
-from namex.resources.auto_analyse.issues import IncorrectCategoryIssue
-
-
-class BcIncorrectCategoryIssue(IncorrectCategoryIssue):
-    pass
