--- api/namex/resources/auto_analyse/paths/bc_name_analysis/issues/name_requires_consent.py
@@ -1,5 +1,0 @@
-from namex.resources.auto_analyse.issues import NameRequiresConsentIssue
-
-
-class BcNameRequiresConsentIssue(NameRequiresConsentIssue):
-    pass
