--- api/namex/resources/auto_analyse/paths/bc_name_analysis/issues/too_many_words.py
@@ -1,5 +1,0 @@
-from namex.resources.auto_analyse.issues import TooManyWordsIssue
-
-
-class BcTooManyWordsIssue(TooManyWordsIssue):
-    pass
