--- api/namex/resources/auto_analyse/paths/bc_name_analysis/issues/word_special_use.py
@@ -1,5 +1,0 @@
-from namex.resources.auto_analyse.issues import WordSpecialUseIssue
-
-
-class BcWordSpecialUseIssue(WordSpecialUseIssue):
-    pass
