+++ api/namex/resources/auto_analyse/paths/name_analysis/__init__.py
@@ -1,0 +1,4 @@
+"""Name Analysis exports."""
+
+from .name_analysis import NameAnalysis
+from .name_analysis import api as name_analysis_api
