+++ api/namex/resources/auto_analyse/paths/name_analysis/issues/__init__.py
@@ -1,0 +1,19 @@
+"""Issues exports."""
+
+from .add_descriptive_word import BcAddDescriptiveWordIssue, XproAddDescriptiveWordIssue
+from .add_distinctive_word import BcAddDistinctiveWordIssue, XproAddDistinctiveWordIssue
+from .check_is_valid import BcCheckIsValidIssue, XproCheckIsValidIssue
+from .contains_unclassifiable_word import BcContainsUnclassifiableWordIssue, XproContainsUnclassifiableWordIssue
+from .contains_words_to_avoid import BcContainsWordsToAvoidIssue, XproContainsWordsToAvoidIssue
+from .corporate_name_conflict import BcCorporateNameConflictIssue, XproCorporateNameConflictIssue
+from .designation_mismatch import BcDesignationMismatchIssue, XproDesignationMismatchIssue
+from .designation_misplaced import BcDesignationMisplacedIssue, XproDesignationMisplacedIssue
+from .designation_non_existent import BcDesignationNonExistentIssue, XproDesignationNonExistentIssue
+from .designation_removal import BcDesignationRemovalIssue
+from .end_designation_more_than_once import BcEndDesignationMoreThanOnceIssue
+from .incorrect_category import BcIncorrectCategoryIssue, XproIncorrectCategoryIssue
+from .incorrect_year import BcIncorrectYearIssue
+from .name_requires_consent import BcNameRequiresConsentIssue, XproNameRequiresConsentIssue
+from .queue_name_conflict import BcQueueNameConflictIssue
+from .too_many_words import BcTooManyWordsIssue, XproTooManyWordsIssue
+from .word_special_use import BcWordSpecialUseIssue, XproWordSpecialUseIssue
