+++ api/namex/resources/auto_analyse/paths/name_analysis/issues/add_descriptive_word.py
@@ -1,14 +1,23 @@
-# Import DTOs
+"""Descriptive issue classes."""
+from namex.resources.auto_analyse.issues import AddDescriptiveWordIssue
+
 from ....response_objects import NameAnalysisIssue
 
-from namex.resources.auto_analyse.issues import DesignationMismatchIssue
+
+class BcAddDescriptiveWordIssue(AddDescriptiveWordIssue):
+    """BC descriptive word issue."""
+
+    pass
 
 
-class XproDesignationMismatchIssue(DesignationMismatchIssue):
+class XproAddDescriptiveWordIssue(AddDescriptiveWordIssue):
+    """XPRO descriptive word issue."""
+
     def create_issue(self):
+        """Create descriptive word issue."""
         issue = NameAnalysisIssue(
             issue_type=self.issue_type,
-            line1="",
+            line1='',
             line2=None,
             consenting_body=None,
             designations=None,
