+++ api/namex/resources/auto_analyse/paths/name_analysis/issues/add_distinctive_word.py
@@ -1,14 +1,23 @@
-# Import DTOs
+"""Distinctive issue classes."""
+from namex.resources.auto_analyse.issues import AddDistinctiveWordIssue
+
 from ....response_objects import NameAnalysisIssue
 
-from namex.resources.auto_analyse.issues import AddDescriptiveWordIssue
+
+class BcAddDistinctiveWordIssue(AddDistinctiveWordIssue):
+    """BC distinctive word issue."""
+
+    pass
 
 
-class XproAddDescriptiveWordIssue(AddDescriptiveWordIssue):
+class XproAddDistinctiveWordIssue(AddDistinctiveWordIssue):
+    """XPRO distinctive word issue."""
+
     def create_issue(self):
+        """Create distinctive word issue."""
         issue = NameAnalysisIssue(
             issue_type=self.issue_type,
-            line1="",
+            line1='',
             line2=None,
             consenting_body=None,
             designations=None,
