+++ api/namex/resources/auto_analyse/paths/name_analysis/issues/contains_words_to_avoid.py
@@ -1,15 +1,24 @@
-# Import DTOs
+"""Contains words to avoid issue classes."""
+from namex.resources.auto_analyse.issues import ContainsWordsToAvoidIssue
+
 from ....response_objects import NameAnalysisIssue
 
-from namex.resources.auto_analyse.issues import DesignationNonExistentIssue
+
+class BcContainsWordsToAvoidIssue(ContainsWordsToAvoidIssue):
+    """BC contains words to avoid."""
+
+    pass
 
 
-class XproDesignationNonExistentIssue(DesignationNonExistentIssue):
+class XproContainsWordsToAvoidIssue(ContainsWordsToAvoidIssue):
+    """XPRO contains words to avoid."""
+
     def create_issue(self):
+        """Create XPRO contains words to avoid issue."""
         issue = NameAnalysisIssue(
             issue_type=self.issue_type,
-            line1="",
-            line2=None,
+            line1='',
+            line2='',
             consenting_body=None,
             designations=None,
             show_reserve_button=False,
