+++ api/namex/resources/auto_analyse/paths/name_analysis/issues/corporate_name_conflict.py
@@ -1,14 +1,23 @@
-# Import DTOs
+"""Corporate Name Conflict issue classes."""
+from namex.resources.auto_analyse.issues import CorporateNameConflictIssue
+
 from ....response_objects import NameAnalysisIssue
 
-from namex.resources.auto_analyse.issues import CorporateNameConflictIssue
+
+class BcCorporateNameConflictIssue(CorporateNameConflictIssue):
+    """BC corporate name conflict issue."""
+
+    pass
 
 
 class XproCorporateNameConflictIssue(CorporateNameConflictIssue):
+    """XPRO corporate name conflict issue."""
+
     def create_issue(self):
+        """Create XPRO corporate name conflict issue."""
         issue = NameAnalysisIssue(
             issue_type=self.issue_type,
-            line1="",
+            line1='',
             line2=None,
             consenting_body=None,
             designations=None,
