+++ api/namex/resources/auto_analyse/paths/name_analysis/issues/designation_removal.py
@@ -1,5 +1,8 @@
+"""Designation removal issue classes."""
 from namex.resources.auto_analyse.issues import DesignationRemovalIssue
 
 
 class BcDesignationRemovalIssue(DesignationRemovalIssue):
+    """BC designation removal issue."""
+
     pass
