+++ api/namex/resources/auto_analyse/paths/name_analysis/issues/end_designation_more_than_once.py
@@ -1,5 +1,8 @@
+"""Multiple end designation issue classes."""
 from namex.resources.auto_analyse.issues import EndDesignationMoreThanOnceIssue
 
 
 class BcEndDesignationMoreThanOnceIssue(EndDesignationMoreThanOnceIssue):
+    """BC multiple end designation issue."""
+
     pass
