+++ api/namex/resources/auto_analyse/paths/name_analysis/issues/incorrect_year.py
@@ -1,5 +1,8 @@
+"""Incorrect Year issue classes."""
 from namex.resources.auto_analyse.issues import IncorrectYearIssue
 
 
 class BcIncorrectYearIssue(IncorrectYearIssue):
+    """BC incorrect year issue."""
+
     pass
