+++ api/namex/resources/auto_analyse/paths/name_analysis/issues/queue_name_conflict.py
@@ -1,5 +1,8 @@
+"""Corporate Name Conflict Issue classes."""
 from namex.resources.auto_analyse.issues import CorporateNameConflictIssue
 
 
 class BcQueueNameConflictIssue(CorporateNameConflictIssue):
+    """BC corporate name conflict issue."""
+
     pass
