--- api/namex/resources/auto_analyse/paths/xpro_name_analysis/__init__.py
@@ -1,2 +1,0 @@
-from .xpro_name_analysis import XproNameAnalysis
-from .xpro_name_analysis import api as xpro_name_analysis_api
