--- api/namex/resources/auto_analyse/paths/xpro_name_analysis/issues/__init__.py
@@ -1,14 +1,0 @@
-from .check_is_valid import XproCheckIsValidIssue
-from .add_descriptive_word import XproAddDescriptiveWordIssue
-from .add_distinctive_word import XproAddDistinctiveWordIssue
-from .check_is_valid import XproCheckIsValidIssue as CheckIsValidIssue
-from .contains_unclassifiable_word import XproContainsUnclassifiableWordIssue
-from .contains_words_to_avoid import XproContainsWordsToAvoidIssue
-from .corporate_name_conflict import XproCorporateNameConflictIssue
-from .designation_mismatch import XproDesignationMismatchIssue
-from .designation_misplaced import XproDesignationMisplacedIssue
-from .designation_non_existent import XproDesignationNonExistentIssue
-from .incorrect_category import XproIncorrectCategoryIssue as XproIncorrectCategoryIssue
-from .name_requires_consent import XproNameRequiresConsentIssue
-from .too_many_words import XproTooManyWordsIssue
-from .word_special_use import XproWordSpecialUseIssue as XproWordSpecialUseIssue
