--- api/namex/resources/auto_analyse/paths/xpro_name_analysis/xpro_name_analysis_issues.py
@@ -1,1 +1,0 @@
-from ...issues import *
