+++ api/namex/services/name_request/auto_analyse/protected_name_analysis.py
@@ -67,6 +67,7 @@ class ProtectedNameAnalysisService(NameAnalysisDirector, SetDesignationsListsMix
 
             # Return any combination of these checks
             if not self.skip_search_conflicts and auto_analyze_config in ('EXACT_MATCH', 'SEARCH_CONFLICTS'):
+                print('do_analysis')
                 check_conflicts = builder.search_exact_match(self.get_list_dist(), self.get_list_desc(),
                                                              self.compound_descriptive_name_tokens,
                                                              False, self.get_designation_end_list_all(),
@@ -108,12 +109,12 @@ class ProtectedNameAnalysisService(NameAnalysisDirector, SetDesignationsListsMix
 
             # TODO: Use the list_name array, don't use a string in the method!
             # check_words_requiring_consent = builder.check_words_requiring_consent(list_name)  # This is correct
-            check_words_requiring_consent = builder.check_words_requiring_consent(
-                self.name_tokens, self.processed_name
-            )
+            # check_words_requiring_consent = builder.check_words_requiring_consent(
+            #     self.name_tokens, self.processed_name
+            # )
 
-            if not check_words_requiring_consent.is_valid:
-                results.append(check_words_requiring_consent)
+            # if not check_words_requiring_consent.is_valid:
+            #     results.append(check_words_requiring_consent)
 
             check_designation_existence = builder.check_designation_existence(
                 self.get_original_name_tokenized(),
@@ -143,12 +144,12 @@ class ProtectedNameAnalysisService(NameAnalysisDirector, SetDesignationsListsMix
                 if not check_designation_misplaced.is_valid:
                     results.append(check_designation_misplaced)
 
-            check_special_words = builder.check_word_special_use(
-                self.name_tokens,
-                self.get_processed_name()
-            )
+            # check_special_words = builder.check_word_special_use(
+            #     self.name_tokens,
+            #     self.get_processed_name()
+            # )
 
-            if not check_special_words.is_valid:
-                results.append(check_special_words)
+            # if not check_special_words.is_valid:
+            #     results.append(check_special_words)
 
         return results
