+++ api/setup.cfg
@@ -1,6 +1,49 @@
+[flake8]
+exclude = .git,*migrations*
+max-line-length = 120
+docstring-min-length=10
+per-file-ignores =
+    */__init__.py:F401,B902
+    # ignore "remove bad quotes" error in tests, since we like to copy-paste json which has double quotes instead of single.
+    */test_*.py:Q000
+    *.py:B902
+
 [pycodestyle]
 max_line_length = 120
 ignore = E501
+docstring-min-length=10
+notes=FIXME,XXX # TODO is ignored
+ignored-modules=flask_sqlalchemy
+    sqlalchemy
+per-file-ignores =
+    */__init__.py:F401
+good-names=
+    b,
+    d,
+    i,
+    e,
+    f,
+    u,
+    k,
+    v,
+    rv,
+    logger,
+
+[pylint]
+ignore=migrations,test
+max_line_length=120
+notes=FIXME,XXX,TODO
+ignored-modules=flask_sqlalchemy,sqlalchemy,SQLAlchemy,alembic,scoped_session
+ignored-classes=scoped_session
+min-similarity-lines=100
+disable=W0511
+
+[isort]
+line_length = 120
+indent = 4
+multi_line_output = 3
+lines_after_imports = 2
+include_trailing_comma = True
 
 [aliases]
 test=pytest
