+++ services/auto-analyze/requirements.txt
@@ -1,41 +1,42 @@
-Flask-SQLAlchemy==2.4.1
+Flask-SQLAlchemy==2.5.1
 Flask==1.1.2
 Hypercorn==0.11.2
 Jinja2==2.11.3
 MarkupSafe==1.1.1
 Quart==0.10.0
-SQLAlchemy==1.4.4
-Werkzeug==0.16.1
-aiofiles==0.6.0
+SQLAlchemy==1.4.11
+Werkzeug==1.0.1
+aiofiles==0.7.0
 blinker==1.4
 certifi==2020.12.5
 chardet==4.0.0
 click==7.1.2
-flask-marshmallow==0.11.0
+flask-marshmallow==0.14.0
 greenlet==1.0.0
 h11==0.12.0
 h2==4.0.0
 hpack==4.0.0
-hyperframe==6.0.0
+hyperframe==6.0.1
 idna==2.10
 inflect==5.3.0
 itsdangerous==1.1.0
+joblib==1.0.1
 jsonpickle==2.0.0
 lxml==4.6.3
-marshmallow-sqlalchemy==0.19.0
-marshmallow==2.19.2
-multidict==5.1.0
-nltk==3.4.5
+marshmallow-sqlalchemy==0.24.3
+marshmallow==3.11.1
+nltk==3.6.2
 numpy==1.20.2
-pandas==1.2.3
+pandas==1.2.4
 priority==1.3.0
 pysolr==3.9.0
 python-dateutil==2.8.1
 pytz==2021.1
+regex==2021.4.4
 requests==2.25.1
 six==1.15.0
-sortedcontainers==2.3.0
 toml==0.10.2
+tqdm==4.60.0
 urllib3==1.26.4
 wsproto==1.0.0
 git+https://github.com/bcgov/namex.git#egg=namex&subdirectory=api
