+++ services/auto-analyze/requirements/dev.txt
@@ -1,5 +1,4 @@
-# Everything the developer needs in addition to the production requirements
--r prod.txt
+# Everything the developer needs other than production requirements
 
 # Testing
 coverage
