+++ services/auto-analyze/requirements/prod.txt
@@ -1,14 +1,14 @@
 hypercorn
 quart
 requests
-nltk==3.4.5
-Flask-SQLAlchemy==2.4.1
-Flask-Marshmallow==0.11.0
-marshmallow==2.19.2
-marshmallow-sqlalchemy==0.19.0
+nltk
+Flask-SQLAlchemy
+Flask-Marshmallow
+marshmallow
+marshmallow-sqlalchemy
 pandas
 inflect
-werkzeug==0.16.1
+werkzeug
 pysolr
 jsonpickle
 lxml
